Перевёл расчёты в наносекунды. Не работает на малых объёмах данных.

+++ copy_between_streams_with_speed/src/main/java/com/github/aakumykov/copy_between_streams_with_speed/percentOf.kt
@@ -3,9 +3,9 @@ package com.github.aakumykov.copy_between_streams_with_speed
 import com.github.aakumykov.copy_between_streams_with_speed.ext.roundToFloatingDigits
 
 fun percentOf(actual: Long, base: Long, digitsAfterComma: Int = 3): String {
-    return percentOf(actual.toDouble(), base.toDouble(), digitsAfterComma)
+    return percentOf(actual.toFloat(), base.toFloat(), digitsAfterComma)
 }
 
-fun percentOf(actual: Double, base: Double, digitsAfterComma: Int = 3): String {
+fun percentOf(actual: Float, base: Float, digitsAfterComma: Int = 3): String {
     return ((actual / base)*100).roundToFloatingDigits(digitsAfterComma)
 }
